snackShackFoodDispenserCode/requirements.txt
flask
gpiozero
RPI.GPIO
opencv-python
requests
